# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Volleyball Tryout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coaches:
    get:
      operationId: listCoaches
      tags: [coaches]
      summary: List all coaches
      responses:
        "200":
          description: List of coaches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coach"
    post:
      operationId: createCoach
      tags: [coaches]
      summary: Create a coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCoachBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
  /coaches/import:
    post:
      operationId: importCoachesCsv
      tags: [coaches]
      summary: Import coaches from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /coaches/{id}:
    delete:
      operationId: deleteCoach
      tags: [coaches]
      summary: Delete a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coaches/{id}/draft:
    get:
      operationId: getCoachDraft
      tags: [coaches]
      summary: Get or create a coach's draft roster
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coach draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachDraft"
  /coaches/{id}/draft/players:
    post:
      operationId: addPlayerToDraft
      tags: [coaches]
      summary: Add a player to a coach's draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftPlayerBody"
      responses:
        "201":
          description: Added
  /coaches/{id}/draft/players/{playerId}:
    delete:
      operationId: removePlayerFromDraft
      tags: [coaches]
      summary: Remove a player from a coach's draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
    patch:
      operationId: togglePlayerLock
      tags: [coaches]
      summary: Toggle lock status of a player on a coach's draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerLockInput"
      responses:
        "200":
          description: Updated
...
  /coaches/{id}/wishlist:
    get:
      operationId: getCoachWishlist
      tags: [coaches]
      summary: Get a coach's wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wishlist player IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
    post:
      operationId: addToWishlist
      tags: [coaches]
      summary: Add a player to a coach's wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
  /coaches/{id}/wishlist/{playerId}:
    delete:
      operationId: removeFromWishlist
      tags: [coaches]
      summary: Remove a player from a coach's wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /ai/player-summary/{playerId}:
    post:
      operationId: generatePlayerSummary
      tags: [ai]
      summary: Generate AI summary for a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI-generated player summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPlayerSummary"
        "404":
          description: Not found
  /ai/roster-explain/{rosterId}:
    post:
      operationId: generateRosterExplanation
      tags: [ai]
      summary: Generate AI explanation for roster decisions
      parameters:
        - name: rosterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI-generated roster explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRosterExplanation"
        "404":
          description: Not found
components:
  schemas:
...
    Player:
      type: object
      required: [id, jerseyNumber, name, position, checkedIn, createdAt, updatedAt]
      properties:
        id:
          type: integer
        jerseyNumber:
          type: string
        name:
          type: string
        position:
          type: string
          enum: [Setter, OutsideHitter, MiddleBlocker, Opposite, Libero]
        checkedIn:
          type: boolean
        heightInches:
          type: ["number", "null"]
        standingReachInches:
          type: ["number", "null"]
        verticalJumpInches:
          type: ["number", "null"]
        overallScore:
          type: ["number", "null"]
        positionScore:
          type: ["number", "null"]
        potentialScore:
          type: ["number", "null"]
        physicalScore:
          type: ["number", "null"]
          description: "Percentile-based score (0–10) from height, reach, and vertical"
        confidenceScore:
          type: ["number", "null"]
          description: "Data completeness score (0–10): how much eval data is available"
        flags:
          type: ["array", "null"]
          items:
            type: string
          description: "Auto-generated player flags (e.g. High Potential, Raw Athlete)"
        rankOverall:
          type: ["integer", "null"]
        rankPosition:
          type: ["integer", "null"]
        rankLocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CsvImport:
      type: object
      required: [csvData]
      properties:
        csvData:
          type: string
    ImportResult:
      type: object
      required: [imported, updated, errors]
      properties:
        imported:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Coach:
      type: object
      required: [id, name, teamName, draftPriority, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        teamName:
          type: string
        draftPriority:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    NewCoachBody:
      type: object
      required: [name, teamName]
      properties:
        name:
          type: string
        teamName:
          type: string
        draftPriority:
          type: array
          items:
            type: string
...
    PlayerLockInput:
      type: object
      required: [locked]
      properties:
        locked:
          type: boolean
    WishlistInput:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: integer
    DraftPlayerBody:
      type: object
      required: [playerId, position]
      properties:
        playerId:
          type: integer
        position:
          type: string
    CoachDraft:
      type: object
      required: [roster, coach, players]
      properties:
        roster:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            coachId:
              type: ["integer", "null"]
        coach:
          $ref: "#/components/schemas/Coach"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
...
    AiPlayerSummary:
      type: object
      required: [playerId, summary, strengths, weaknesses, risks, positionFit, potentialNote]
      properties:
        playerId:
          type: integer
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        positionFit:
          type: string
        potentialNote:
          type: string
        suggestedPositionChange:
          type: ["string", "null"]
    AiRosterExplanation:
      type: object
      required: [rosterId, explanation, playerDecisions]
      properties:
        rosterId:
          type: integer
        explanation:
          type: string
        playerDecisions:
          type: array
          items:
            $ref: "#/components/schemas/PlayerDecision"
    PlayerDecision:
      type: object
      required: [playerId, playerName, decision, reason]
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        decision:
          type: string
          enum: [selected, bubble, excluded]
        reason:
          type: string