# lib/api-zod/src/generated/types/aiPlayerSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Volleyball Tryout Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AiPlayerSummary {
  playerId: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  risks: string[];
  positionFit: string;
  potentialNote: string;
  /** @nullable */
  suggestedPositionChange?: string | null;
}